requirements.txt
absl-py==0.7.1
astor==0.7.1
bert-serving-client==1.8.9
bert-serving-server==1.8.9
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.3.2
gast==0.2.2
GPUtil==1.4.0
grpcio==1.20.1
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.3
protobuf==3.7.1
python-engineio==3.5.1
python-socketio==3.1.2
pyzmq==18.0.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
